+++ astropy_helpers/tests/test_setup_helpers.py
@@ -2,6 +2,7 @@ import os
 import sys
 import stat
 import shutil
+import warnings
 import contextlib
 
 import pytest
@@ -12,9 +13,14 @@ from setuptools import Distribution
 
 from ..setup_helpers import get_package_info, register_commands
 from ..commands import build_ext
+from ..utils import AstropyDeprecationWarning
 
 from . import reset_setup_helpers, reset_distutils_log, fix_hide_setuptools  # noqa
 from . import run_setup, cleanup_import
+
+
+# Determine whether we're in a PY2 environment without using six
+USING_PY2 = sys.version_info < (3,0,0)
 
 
 def _extension_test_package(tmpdir, request, extension_type='c'):
@@ -170,11 +176,18 @@ def test_compiler_module(c_extension_test_package):
     with test_pkg.as_cwd():
         # This is one of the simplest ways to install just a package into a
         # test directory
-        run_setup('setup.py',
-                  ['install',
-                   '--single-version-externally-managed',
-                   '--install-lib={0}'.format(install_temp),
-                   '--record={0}'.format(install_temp.join('record.txt'))])
+        with warnings.catch_warnings(record=True) as w:
+            run_setup('setup.py',
+                      ['install',
+                       '--single-version-externally-managed',
+                       '--install-lib={0}'.format(install_temp),
+                       '--record={0}'.format(install_temp.join('record.txt'))])
+        # Skip this portion of the test on windows systems with Py 2.7 since
+        # it is known to produce additional warnings.
+        if not (USING_PY2 or sys.platform.startswith('win')):
+            # Warning expected from get_git_devstr, called by generate_version_py
+            assert len(w) == 1
+            assert str(w[0].message).startswith("No git repository present at")
 
     with install_temp.as_cwd():
         import apyhtest_eva
@@ -205,7 +218,12 @@ def test_no_cython_buildext(c_extension_test_package, monkeypatch):
                         lambda *args: False)
 
     with test_pkg.as_cwd():
-        run_setup('setup.py', ['build_ext', '--inplace'])
+        with warnings.catch_warnings(record=True) as w:
+            run_setup('setup.py', ['build_ext', '--inplace'])
+        # Warning expected from get_git_devstr, called by generate_version_py
+        if not USING_PY2:
+            assert len(w) == 1
+            assert str(w[0].message).startswith("No git repository present at")
 
     sys.path.insert(0, str(test_pkg))
 
@@ -234,7 +252,13 @@ def test_missing_cython_c_files(pyx_extension_test_package, monkeypatch):
 
     with test_pkg.as_cwd():
         with pytest.raises(SystemExit) as exc_info:
-            run_setup('setup.py', ['build_ext', '--inplace'])
+            with warnings.catch_warnings(record=True) as w:
+                run_setup('setup.py', ['build_ext', '--inplace'])
+            # Warning expected from get_git_devstr, called by generate_version_py
+            if not USING_PY2:
+                assert len(w) == 1
+                assert str(w[0].message).startswith(
+                    "No git repository present at")
 
     msg = ('Could not find C/C++ file '
            '{0}.(c/cpp)'.format('apyhtest_eva/unit02'.replace('/', os.sep)))
@@ -323,7 +347,8 @@ def test_build_docs(tmpdir, mode):
         elif mode == 'cli-l':
             run_setup('setup.py', ['build_docs', '-l'])
         elif mode == 'deprecated':
-            run_setup('setup.py', ['build_sphinx'])
+            with pytest.warns(AstropyDeprecationWarning):
+                run_setup('setup.py', ['build_sphinx'])
         elif mode == 'direct':  # to check coverage
             with docs_dir.as_cwd():
                 from sphinx import main